[aur] Consolidate modules
